Extract datasource config package
