+++ pkg/config/config_test.go
@@ -1,0 +1,81 @@
+package config
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestResolveBaseURL(t *testing.T) {
+	tests := []struct {
+		name     string
+		apiHost  string
+		edge     string
+		edgeURL  string
+		expected string
+	}{
+		{
+			name:     "no edge - uses apiHost",
+			apiHost:  "https://api.axiom.co",
+			expected: "https://api.axiom.co",
+		},
+		{
+			name:     "edge domain",
+			apiHost:  "https://api.axiom.co",
+			edge:     "eu-central-1.aws.edge.axiom.co",
+			expected: "https://eu-central-1.aws.edge.axiom.co",
+		},
+		{
+			name:     "edge domain with trailing slash",
+			apiHost:  "https://api.axiom.co",
+			edge:     "us-east-1.aws.edge.axiom.co/",
+			expected: "https://us-east-1.aws.edge.axiom.co",
+		},
+		{
+			name:     "edgeURL without path",
+			apiHost:  "https://api.axiom.co",
+			edgeURL:  "https://eu-central-1.aws.edge.axiom.co",
+			expected: "https://eu-central-1.aws.edge.axiom.co",
+		},
+		{
+			name:     "edgeURL with trailing slash",
+			apiHost:  "https://api.axiom.co",
+			edgeURL:  "https://eu-central-1.aws.edge.axiom.co/",
+			expected: "https://eu-central-1.aws.edge.axiom.co",
+		},
+		{
+			name:     "edgeURL with custom path - used as-is",
+			edgeURL:  "http://localhost:3400/query",
+			expected: "http://localhost:3400/query",
+		},
+		{
+			name:     "edgeURL takes precedence over edge domain",
+			edgeURL:  "https://primary.edge.axiom.co",
+			edge:     "secondary.edge.axiom.co",
+			expected: "https://primary.edge.axiom.co",
+		},
+		{
+			name:     "legacy EU instance - no edge",
+			apiHost:  "https://api.eu.axiom.co",
+			expected: "https://api.eu.axiom.co",
+		},
+		{
+			name:     "staging edge domain",
+			edge:     "us-east-1.edge.staging.axiomdomain.co",
+			expected: "https://us-east-1.edge.staging.axiomdomain.co",
+		},
+		{
+			name:     "no apiHost, no edge - uses default cloud endpoint",
+			expected: "https://api.axiom.co",
+		},
+	}
+
+	for _, test := range tests {
+		test := test
+		t.Run(test.name, func(t *testing.T) {
+			endpoint, err := resolveEdgeUrl(test.apiHost, test.edge, test.edgeURL)
+			require.NoError(t, err)
+			require.Equal(t, test.expected, endpoint)
+		})
+	}
+}
