+++ pkg/plugin/datasource_test.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/axiomhq/axiom-go/axiom/query"
 	"github.com/axiomhq/axiom-grafana/pkg/axiomapi"
+	"github.com/axiomhq/axiom-grafana/pkg/config"
 	"github.com/grafana/grafana-plugin-sdk-go/data"
 	"github.com/stretchr/testify/require"
 
@@ -39,84 +40,6 @@ func TestQueryData(t *testing.T) {
 	require.Len(t, resp.Responses, 3, "QueryData must return a response for each query")
 }
 
-func TestResolveBaseURL(t *testing.T) {
-	tests := []struct {
-		name     string
-		apiHost  string
-		edge     string
-		edgeURL  string
-		expected string
-	}{
-		{
-			name:     "no edge - uses apiHost",
-			apiHost:  "https://api.axiom.co",
-			expected: "https://api.axiom.co",
-		},
-		{
-			name:     "edge domain",
-			apiHost:  "https://api.axiom.co",
-			edge:     "eu-central-1.aws.edge.axiom.co",
-			expected: "https://eu-central-1.aws.edge.axiom.co",
-		},
-		{
-			name:     "edge domain with trailing slash",
-			apiHost:  "https://api.axiom.co",
-			edge:     "us-east-1.aws.edge.axiom.co/",
-			expected: "https://us-east-1.aws.edge.axiom.co",
-		},
-		{
-			name:     "edgeURL without path",
-			apiHost:  "https://api.axiom.co",
-			edgeURL:  "https://eu-central-1.aws.edge.axiom.co",
-			expected: "https://eu-central-1.aws.edge.axiom.co",
-		},
-		{
-			name:     "edgeURL with trailing slash",
-			apiHost:  "https://api.axiom.co",
-			edgeURL:  "https://eu-central-1.aws.edge.axiom.co/",
-			expected: "https://eu-central-1.aws.edge.axiom.co",
-		},
-		{
-			name:     "edgeURL with custom path - used as-is",
-			edgeURL:  "http://localhost:3400/query",
-			expected: "http://localhost:3400/query",
-		},
-		{
-			name:     "edgeURL takes precedence over edge domain",
-			edgeURL:  "https://primary.edge.axiom.co",
-			edge:     "secondary.edge.axiom.co",
-			expected: "https://primary.edge.axiom.co",
-		},
-		{
-			name:     "legacy EU instance - no edge",
-			apiHost:  "https://api.eu.axiom.co",
-			expected: "https://api.eu.axiom.co",
-		},
-		{
-			name:     "staging edge domain",
-			edge:     "us-east-1.edge.staging.axiomdomain.co",
-			expected: "https://us-east-1.edge.staging.axiomdomain.co",
-		},
-		{
-			name:     "no apiHost, no edge - uses default cloud endpoint",
-			expected: "https://api.axiom.co",
-		},
-	}
-
-	for _, test := range tests {
-		test := test
-		t.Run(test.name, func(t *testing.T) {
-			endpoint, err := resolveBaseUrl(urlInput{
-				APIHost: test.apiHost,
-				Edge:    test.edge,
-				EdgeURL: test.edgeURL,
-			})
-			require.NoError(t, err)
-			require.Equal(t, test.expected, endpoint)
-		})
-	}
-}
-
 func TestResourceHandlerFetchesEscapedMetricAutocompleteValues(t *testing.T) {
 	upstream := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Content-Type", "application/json")
@@ -135,8 +58,8 @@ func TestResourceHandlerFetchesEscapedMetricAutocompleteValues(t *testing.T) {
 	defer upstream.Close()
 
 	ds := Datasource{
-		api: axiomapi.NewClient(axiomapi.Config{
-			APIURL:  upstream.URL,
+		api: axiomapi.NewClient(&config.PluginConfig{
+			APIHost: upstream.URL,
 			EdgeURL: upstream.URL,
 		}),
 	}
@@ -219,8 +142,8 @@ func TestQueryLogsVolumeReturnsFullRangeHistogramFrame(t *testing.T) {
 
 	queryText := "['logs'] | where level == 'error'"
 	ds := Datasource{
-		api: axiomapi.NewClient(axiomapi.Config{
-			APIURL:  upstream.URL,
+		api: axiomapi.NewClient(&config.PluginConfig{
+			APIHost: upstream.URL,
 			EdgeURL: upstream.URL,
 		}),
 	}
