+++ pkg/util/strings.go
@@ -1,0 +1,8 @@
+package util
+
+func CheckString(i interface{}) string {
+	if str, ok := i.(string); ok {
+		return str
+	}
+	return ""
+}
