+++ pkg/version/version.go
@@ -1,4 +1,4 @@
-package plugin
+package version
 
 // TODO: automate parsing the version from the package.json file
-const Version = "0.6.4"
+const Version = "0.7.0"
